+++ useScroll/useScroll.js
@@ -369,7 +369,7 @@ const useScroll = (props) => {
 		...rest,
 		...scrollProps,
 		assignProperties,
-		noScrollByDrag: !platform.touchscreen,
+		noScrollByDrag: !platform.touchScreen,
 		addEventListeners,
 		handleResizeWindow,
 		horizontalScrollbarHandle,